+++ base/shared/src/main/scala/scalaz/Prelude.scala
@@ -13,7 +13,7 @@ trait Prelude  extends data.DisjunctionFunctions
                   with typeclass.BindFunctions
                   with typeclass.FunctorFunctions
                   with typeclass.InvariantFunctorFunctions
-                  with typeclass.AsFunctions
+                  with AsFunctions
                   with typeclass.TraversableFunctions {
 
   // Base Class
@@ -101,8 +101,8 @@ trait Prelude  extends data.DisjunctionFunctions
   
   type \/[L, R] = data.Disjunction.\/[L, R]
   type ===[A, B] = data.Is[A, B]
-  type <~<[-A, +B] = typeclass.As[A, B]
-  type >~>[+B, -A] = typeclass.As[A, B]
+  type <~<[-A, +B] = As[A, B]
+  type >~>[+B, -A] = As[A, B]
   type Identity[A] = data.Identity[A]
   type Forget[A, B, C] = data.Forget[A, B, C]
