+++ base/shared/src/main/scala/scalaz/data/AsFunctions.scala
@@ -1,7 +1,8 @@
 package scalaz
-package typeclass
+package data
 
 import Prelude.<~<
+import typeclass.{IsContravariant, IsCovariant}
 
 trait AsFunctions {
   /**Lift Scala's subtyping relationship */
