+++ base/shared/src/main/scala/scalaz/data/Forall.scala
@@ -3,7 +3,6 @@ package data
 
 import Prelude.<~<
 
-import typeclass.As
 import scala.language.implicitConversions
 
 trait ForallModule {
