+++ base/shared/src/main/scala/scalaz/typeclass/IsCovariantClass.scala
@@ -2,6 +2,7 @@ package scalaz
 package typeclass
 
 import Prelude.<~<
+import data.As
 
 trait IsCovariantClass[F[_]] extends IsCovariant[F] {
   final def isCovariant: IsCovariant[F] = this
